Refactor: add GetAlignedCommands for SRV-sorted SubUV batching, fix C4244 warning
PrepareBatchers now routes commands through GetAlignedCommands, which sorts
SubUV pass by Particle pointer so same-SRV sprites are contiguous.
Also fix double-to-float narrowing warning in PrimitiveComponent AABB calc.

+++ W3_Jungle_Team6/Source/Engine/Component/PrimitiveComponent.cpp
@@ -169,17 +169,17 @@ void USphereComponent::UpdateWorldAABB() const
 {
 	FVector Center = { CachedWorldMatrix.M[3][0], CachedWorldMatrix.M[3][1], CachedWorldMatrix.M[3][2] };
 
-	float ExtentX = LocalExtents.X * std::sqrt(std::pow(CachedWorldMatrix.M[0][0], 2) +
-		std::pow(CachedWorldMatrix.M[1][0], 2) +
-		std::pow(CachedWorldMatrix.M[2][0], 2));
-								  
-	float ExtentY = LocalExtents.Y * std::sqrt(std::pow(CachedWorldMatrix.M[0][1], 2) +
-		std::pow(CachedWorldMatrix.M[1][1], 2) +
-		std::pow(CachedWorldMatrix.M[2][1], 2));
-								  
-	float ExtentZ = LocalExtents.Z * std::sqrt(std::pow(CachedWorldMatrix.M[0][2], 2) +
-		std::pow(CachedWorldMatrix.M[1][2], 2) +
-		std::pow(CachedWorldMatrix.M[2][2], 2));
+	float ExtentX = LocalExtents.X * sqrtf(powf(CachedWorldMatrix.M[0][0], 2) +
+		powf(CachedWorldMatrix.M[1][0], 2) +
+		powf(CachedWorldMatrix.M[2][0], 2));
+
+	float ExtentY = LocalExtents.Y * sqrtf(powf(CachedWorldMatrix.M[0][1], 2) +
+		powf(CachedWorldMatrix.M[1][1], 2) +
+		powf(CachedWorldMatrix.M[2][1], 2));
+
+	float ExtentZ = LocalExtents.Z * sqrtf(powf(CachedWorldMatrix.M[0][2], 2) +
+		powf(CachedWorldMatrix.M[1][2], 2) +
+		powf(CachedWorldMatrix.M[2][2], 2));
 
 	WorldAABBMinLocation = { Center.X - ExtentX, Center.Y - ExtentY, Center.Z - ExtentZ };
 	WorldAABBMaxLocation = { Center.X + ExtentX, Center.Y + ExtentY, Center.Z + ExtentZ };
